wip: 구조 설정
