+++ packages/map/src/index.tsx
@@ -1,6 +1,4 @@
 export { default as MapView } from './map-view'
 export * from './focus-tracker'
+export * from './overlay'
 export * from './utilities'
-export * from './marker'
-export * from './polygon'
-export * from './polyline'
