+++ packages/map/src/overlay/marker/index.ts
@@ -1,0 +1,1 @@
+export * from './circle-marker'
