--- packages/map/src/polygon/index.ts
@@ -1,1 +1,0 @@
-export * from './polygon'
